Adicionando o arquivo de cabeçalho no arquivo de funções.c e arrumando o tipo de quantidade na struct Carrinho.

+++ Structs.h
@@ -6,7 +6,7 @@
 
 typedef struct Carrinho {
     long long int codigo_produto; 
-    int quantidade;
+    long long int quantidade;
     struct Carrinho * proximo;
 } Carrinho;
